pyproject.toml
[project]
name = "3ddl-wp5-swip"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "easydict>=1.13",
    "matplotlib>=3.10.8",
    "nibabel>=5.3.3",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "ultralytics>=8.3.240",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E", "W", "F", "C901", "B", "SIM", "Q", "UP"]
ignore = []
unfixable = ["F401"]
